#16674 Fix selected rows export with client-side ordering (#17239)
* #16674 Fix selected rows export with client-side ordering

* #16674 Change ResultSetDataContainerOptions.selectedRows type to Integer

+++ plugins/org.jkiss.dbeaver.ui.editors.data/src/org/jkiss/dbeaver/ui/controls/resultset/handler/ResultSetHandlerCopyAs.java
@@ -105,9 +105,9 @@ public class ResultSetHandlerCopyAs extends AbstractHandler implements IElementU
         List<ResultSetRow> rsSelectedRows = rsSelection.getSelectedRows();
         List<DBDAttributeBinding> rsSelectedAttributes = rsSelection.getSelectedAttributes();
         if (rsSelectedRows.size() > 1 || rsSelectedAttributes.size() > 1) {
-            List<Long> selectedRows = new ArrayList<>();
+            List<Integer> selectedRows = new ArrayList<>();
             for (ResultSetRow selectedRow : rsSelectedRows) {
-                selectedRows.add((long) selectedRow.getRowNumber());
+                selectedRows.add(selectedRow.getRowNumber());
             }
 
             options.setSelectedRows(selectedRows);
